requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
amazon-textract-caller==0.2.2
amazon-textract-response-parser==1.0.2
amazon-textract-textractor==1.7.9
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
apispec==6.6.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
bce-python-sdk==0.9.5
beautifulsoup4==4.12.3
bech32==1.2.0
bleach==6.1.0
blinker==1.7.0
boto3==1.34.74
botocore==1.34.74
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.1
camelot-py==0.9.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
configobj==5.0.8
configparser==6.0.1
contextlib2==21.6.0
contourpy==1.2.0
cosmpy==0.9.2
cryptography==42.0.5
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.9
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
datasets==2.18.0
db-dtypes==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.7
distro==1.9.0
dm-tree==0.1.8
duckdb==0.9.1
ecdsa==0.18.0
editdistance==0.8.1
effdet==0.4.1
Elixir==0.7.1
emoji==2.10.1
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.1
fastapi==0.103.1
fastjsonschema==2.19.1
filelock==3.12.4
filetype==1.2.0
fireworks-ai==0.13.0
fitz==0.0.1.dev2
Flask==3.0.2
flask-babel==4.0.0
flatbuffers==24.3.7
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
future==1.0.0
gast==0.5.4
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.60.0
graphql-core==3.2.3
grpcio==1.62.1
gspread==5.11.3
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
importlib-metadata==7.0.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.2
ipython==8.21.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.0.5
kiwisolver==1.4.5
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.37
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.38
layoutparser==0.3.4
lazy_loader==0.3
libclang==16.0.6
libcst==1.0.1
lmdb==1.4.1
looseversion==1.3.0
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mplcursors==0.5.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.7
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.3
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.15.1
openai==1.16.2
openapi-generator-cli==4.3.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
paddleocr==2.6.0.1
paddlepaddle==2.6.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pathlib==1.0.1
pbr==6.0.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pikepdf==8.14.0
pillow==10.3.0
pillow_heif==0.15.0
platformdirs==3.11.0
pluggy==1.4.0
portalocker==2.8.2
premailer==3.10.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.24.3
prov==2.0.0
psutil==5.9.8
psycopg2==2.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.15
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
pydyf==0.9.0
Pygments==2.16.1
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.1
pypdf==4.1.0
PyPDF2==3.0.1
pypdfium2==4.28.0
pyphen==0.14.0
pytesseract==0.3.10
pytest==8.1.1
python-dateutil==2.8.2
python-doctr==0.8.1
python-iso639==2024.2.7
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.6.1
rarfile==4.1
rdflib==7.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
schema==0.7.5
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
shapely==2.0.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlalchemy-migrate==0.11.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
strawberry-graphql==0.209.2
streamlit==1.31.1
sympy==1.12
tabulate==0.9.0
Tempita==0.5.2
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
traits==6.3.2
transformers==4.38.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uagents==0.11.1
Unidecode==1.3.8
unstructured==0.12.6
unstructured-client==0.18.0
unstructured-inference==0.7.25
unstructured.pytesseract==0.3.12
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.20.0
validators==0.22.0
virtualenv==20.24.5
visualdl==2.5.3
wcwidth==0.2.13
weasyprint==61.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.2
wrapt==1.16.0
xlrd==0.7.1
XlsxWriter==3.2.0
xlwt==0.7.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zopfli==0.2.3
